fix(hostboot): drop redundant int64() conversion on darwin tv.Sec (ga-47doqg)
unix.Timeval.Sec is already int64 on darwin, so wrapping it in int64()
is a no-op that golangci-lint's unconvert rejects under `make lint-full`
(internal/hostboot/hostboot_darwin.go:21:24). This is the sole failure
in the Mac / quality job; every needs-mac job in CI declares
needs: [mac-quality, ...] and is SKIPPED behind it, so this one line
currently blocks all Mac-tier CI signal fleet-wide (e.g. #6192, #6201).

int64(tv.Usec) is left untouched -- Usec is int32 there, so that
conversion is real.

Verified: GOOS=darwin GOARCH=arm64 go build/vet/golangci-lint all
clean on internal/hostboot (0 lint issues, previously 1); native Linux
go vet ./... and internal/hostboot tests pass unchanged (mathematically
identical value, same width, no behavior change).

+++ internal/hostboot/hostboot_darwin.go
@@ -18,5 +18,5 @@ func bootTime() (time.Time, error) {
 	if tv.Sec <= 0 {
 		return time.Time{}, fmt.Errorf("hostboot: implausible kern.boottime seconds %d", tv.Sec)
 	}
-	return time.Unix(int64(tv.Sec), int64(tv.Usec)*1000), nil
+	return time.Unix(tv.Sec, int64(tv.Usec)*1000), nil
 }
